+++ cordapp/build.gradle
@@ -1,5 +1,10 @@
 apply plugin: 'net.corda.plugins.cordapp'
 
+cordapp {
+    targetPlatformVersion 4
+    minimumPlatformVersion 4
+}
+
 dependencies {
     cordaCompile "${corda_release_group}:corda-core:${corda_release_version}"
     cordaCompile "${corda_release_group}:corda-rpc:${corda_release_version}"
