+++ cordlib/src/main/kotlin/com.luxoft.blockchainlab.corda.hyperledger.indy/flow/CreateRevocationRegistryFlow.kt
@@ -10,7 +10,10 @@ import com.luxoft.blockchainlab.hyperledger.indy.models.CredentialDefinitionId
 import com.luxoft.blockchainlab.hyperledger.indy.models.RevocationRegistryInfo
 import net.corda.core.contracts.Command
 import net.corda.core.contracts.StateAndContract
-import net.corda.core.flows.*
+import net.corda.core.flows.FlowException
+import net.corda.core.flows.FlowLogic
+import net.corda.core.flows.InitiatingFlow
+import net.corda.core.flows.StartableByRPC
 import net.corda.core.transactions.TransactionBuilder
 
 /**
@@ -84,7 +87,7 @@ object CreateRevocationRegistryFlow {
 
                 val selfSignedTx = serviceHub.signInitialTransaction(trxBuilder, ourIdentity.owningKey)
 
-                subFlow(FinalityFlow(selfSignedTx))
+                finalizeTransaction(selfSignedTx)
 
                 return revocationRegistryInfo
             } catch (t: Throwable) {
